backend/requirements.txt
pillow>=10.4.0
numpy>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

